building private exchange

+++ contracts/kyc/Election.sol
@@ -2,9 +2,10 @@
 pragma solidity >=0.7.0 <0.9.0;
 
 contract Election {
-    uint public votes;
+    uint256 public votes;
     mapping(address => bool) public hasVoted;
 
+    /// @param voterAddr The address voting on the election
     function vote(address voterAddr) public {
         require(
             hasVoted[voterAddr] == false,
@@ -15,6 +16,7 @@ contract Election {
         hasVoted[voterAddr] = true;
     }
 
+    /// @param voterAddr The address voting on the election
     function removeVote(address voterAddr) public {
         require(
             hasVoted[voterAddr] == true,
